:truck: rename : contentList 폴더명 변경
